New info on return of todo and new conditions

+++ backend/src/todo/todo.service.ts
@@ -4,23 +4,29 @@ import { UpdateTodoDto } from './dto/update-todo.dto';
 import { InjectModel } from '@nestjs/sequelize';
 import { Todo } from './entities/todo.entity';
 import { User } from 'src/users/entities/user.entity';
-import { DeleteTodoDto } from './dto/delete-todo.dto';
 
 @Injectable()
 export class TodoService {
-  constructor(@InjectModel(Todo) private todoModel: typeof Todo,
-  @InjectModel(User) private userModel: typeof User ) {}
+  constructor(@InjectModel(Todo) private todoModel: typeof Todo ) {}
 
 
-  async create(dto: CreateTodoDto) {
+  async create(dto: CreateTodoDto, user) {
     const todo = await this.todoModel.create({
-      title: dto.title
+      title: dto.title, user_id: user.id
     })
 
     return todo
   }
   async findAll() {
-    const todo = await this.todoModel.findAll()
+    const todo = await this.todoModel.findAll({
+      attributes: ["id", "title"],
+      include: [
+        {
+          model: User, as: "user",
+          attributes: ["id", "name", "email"]
+        }
+      ]
+    })
 
     return todo;
   }
@@ -33,18 +39,38 @@ export class TodoService {
     return todo;
   }
 
-  async update(@Param("id") id: number, dto: UpdateTodoDto) {
+  async update(id: number, dto: UpdateTodoDto, user) {
     const todo = await this.todoModel.update({ title: dto.title }, { 
       where: {
-        id 
+        id,
+        user_id: user.id
       }
     })
-    return todo
-  }
 
-  async remove(@Param("id") id: number) {
-    const todo = await this.todoModel.destroy({ where: { id } })
+    if (todo[0] >= 1) { 
+      return { 
+        message: "Todo atualizado com sucesso!"
+      }
+    } else if (todo[0] === 0){
+      return {
+        message: "Algo deu errado. Verifique se esse item existe ou se foi voce quem o criou."
+      }
+    }
+    
+  } // Somente o usuário que cria faz o update, se o usuário que chamou pra fazer o update nao for o usuário que criou o todo, lançar um error
+  // O que acontece se o todo nao existir? Como prevenir o update de um todo que nao existe?
 
-    return todo
-  }
+  async remove(id: number, user) {
+    const todo = await this.todoModel.destroy({ where: { id, user_id: user.id } })
+ 
+    if (todo === 1) {
+      return {
+        message: "Item removido com sucesso."
+      }
+    } else {
+      return {
+        message: "Algo deu errado. Verifique se esse item existe ou se foi voce quem o criou."
+      }
+    } // Somente o usuário que criou pode deletar, O que acontece se o todo nao existir? Como prevenir o remove de um todo que nao existe? E se o usuário for deletado?
+  } 
 }
